throwback-api/requirements.txt
Flask==3.0.3
minio==7.2.7
peewee==3.17.5
Pillow==10.4.0
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
Requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
